OnvifModule: Split object in 2 files: behaviors (states) and functions Moved subscription behaviors to a new package.
